Move tab store to separate file
